refactor : 블로그 관련 타입 분리
